UNIONVMS-1866
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/trunk/Modules/Activity/APP@13933 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/test/java/eu/europa/ec/fisheries/ers/service/util/ActivityDataUtil.java
@@ -115,7 +115,7 @@ public class ActivityDataUtil {
         return sizeDistributionEntity;
     }
 
-    public static FaCatchEntity getFaCatchEntity(FishingActivityEntity fishingActivityEntity,String typeCode, String typeCodeListId, String  speciesCode, String speciesCodeListid,Long unitQuantity,Double weightMeasure,String weightMeasureUnitCode,
+    public static FaCatchEntity getFaCatchEntity(FishingActivityEntity fishingActivityEntity,String typeCode, String typeCodeListId, String  speciesCode, String speciesCodeListid,Double unitQuantity,Double weightMeasure,String weightMeasureUnitCode,
              String weighingMeansCode,String weighingMeansCodeListId,SizeDistributionEntity sizeDistributionEntity  ){
         FaCatchEntity faCatchEntity = new FaCatchEntity();
         faCatchEntity.setFishingActivity(fishingActivityEntity);
